+++ src/communication/serial/serial_communication.py
@@ -1,164 +1,16 @@
-# Import necessary modules at the top
-from abc import ABC, abstractmethod
-from datetime import datetime, timezone
-import time
-from typing import Any
-from pyubx2 import UBXReader, NMEA_PROTOCOL, RTCM3_PROTOCOL, UBX_PROTOCOL
-from serial import SerialException
-import logging
 import asyncio
-from serial import Serial
+import logging
+from typing import Any
+from handlers.gngga_handler import GNGGAHandler
+from handlers.navpvt_handler import NAVPVTHandler
+from serial import Serial, SerialException
 from src.interfaces.communication_interface import CommunicationInterface
+from src.communication.serial.message_processor import MessageProcessor
+from src.communication.serial.message_reader import MessageReader
+from pyubx2 import UBX_PROTOCOL, NMEA_PROTOCOL, RTCM3_PROTOCOL
+
 
 logger = logging.getLogger(__name__)
-
-# Define the base handler interface
-class MessageHandler(ABC):
-    @abstractmethod
-    def process(self, parsed_data, device_id, experiment_id):
-        pass
-
-# Implement a handler for GNGGA messages
-class GNGGAHandler(MessageHandler):
-    def process(self, parsed_data, device_id, experiment_id):
-
-        data_dict = {
-            "message_type": parsed_data.identity,
-            "full_time": datetime.now(tz=timezone.utc).isoformat(),  
-            "lat": parsed_data.lat,
-            "ns": parsed_data.NS,
-            "lon": parsed_data.lon,
-            "ew": parsed_data.EW,
-            "quality": parsed_data.quality,
-            "num_sv": parsed_data.numSV,
-            "hdop": parsed_data.HDOP,
-            "alt": parsed_data.alt,
-            "alt_unit": parsed_data.altUnit,
-            "sep": parsed_data.sep,
-            "sep_unit": parsed_data.sepUnit,
-            "diff_age": parsed_data.diffAge or None,
-            "diff_station": parsed_data.diffStation,
-            "processed_time": f"{int(time.time()*1000)}",
-            "device_id": device_id,
-            "experiment_id": experiment_id
-        }
-        return data_dict
-    
-
-class NAVPVTHandler(MessageHandler):
-    def process(self, parsed_data, device_id, experiment_id):
-        logger.info("Processing NAV-PVT message: %s", parsed_data)
-
-        full_time = datetime.now(tz=timezone.utc).isoformat() 
-        logger.debug(parsed_data)
-        data_dict = {
-            "message_type": "NAV-PVT",
-            "full_time": full_time,
-            "year": parsed_data.year,
-            "month": parsed_data.month,
-            "day": parsed_data.day,
-            "hour": parsed_data.hour,
-            "min": parsed_data.min,
-            "second": parsed_data.second,
-            "validDate": parsed_data.validDate,
-            "validTime": parsed_data.validTime,
-            "tAcc": parsed_data.tAcc,
-            "fixType": parsed_data.fixType,
-            "gnssFixOk": parsed_data.gnssFixOk,
-            "numSV": parsed_data.numSV,
-            "lon": parsed_data.lon,
-            "lat": parsed_data.lat,
-            "height": parsed_data.height,
-            "hMSL": parsed_data.hMSL,
-            "hAcc": parsed_data.hAcc,
-            "vAcc": parsed_data.vAcc,
-            "velN": parsed_data.velN,
-            "velE": parsed_data.velE,
-            "velD": parsed_data.velD,
-            "gSpeed": parsed_data.gSpeed,
-            "headMot": parsed_data.headMot,
-            "sAcc": parsed_data.sAcc,
-            "headAcc": parsed_data.headAcc,
-            "pDOP": parsed_data.pDOP,
-            "device_id": device_id,
-            "experiment_id": experiment_id
-        }
-        return data_dict
-
-class MessageReader:
-    """
-    Responsible for reading messages from a stream.
-    """
-    @staticmethod
-    def read_messages(stream, protfilter):
-        ubx_reader = UBXReader(stream, protfilter=protfilter)
-        while stream.in_waiting:
-            try:
-                _, parsed_data = ubx_reader.read()
-                if parsed_data is None:
-                    break  # No more data available
-                yield parsed_data
-            except Exception as e:
-                logger.error("Error reading GNSS message: %s", e)
-                continue 
-
-class MessageProcessor:
-    """
-    Processes GNSS messages using dynamically registered handlers.
-    """
-    handlers = {}
-
-    @classmethod
-    def register_handler(cls, message_type, handler):
-        """
-        Registers a handler for a specific message type.
-
-        :param message_type: The type of the message the handler can process.
-        :param handler: The handler instance.
-        """
-        if message_type in cls.handlers:
-            logger.warning(f"Handler for {message_type} already registered. Overwriting.")
-        cls.handlers[message_type] = handler
-        logger.info(f"Handler for {message_type} registered successfully.")
-
-    @classmethod
-    def deregister_handler(cls, message_type):
-        """
-        Deregisters the handler for a specific message type.
-
-        :param message_type: The type of the message to deregister the handler for.
-        """
-        if message_type in cls.handlers:
-            del cls.handlers[message_type]
-            logger.info(f"Handler for {message_type} deregistered successfully.")
-        else:
-            logger.warning(f"No handler registered for {message_type}.")
-
-    @classmethod
-    def process_data(cls, stream, device_id, gnss_messages, experiment_id):
-        message_reader = MessageReader()
-        try:
-            for parsed_data in message_reader.read_messages(stream, UBX_PROTOCOL | NMEA_PROTOCOL | RTCM3_PROTOCOL):
-                if parsed_data.identity not in gnss_messages:
-                    logger.debug(f"Message type not in GNSS messages: {parsed_data.identity}")
-                    continue
-
-                handler = cls.handlers.get(parsed_data.identity)
-                if not handler:
-                    logger.warning(f"No handler for message type: {parsed_data.identity}")
-                    continue
-
-                return handler.process(parsed_data, device_id, experiment_id)
-        except SerialException as e:
-            logger.error(f"Error reading data from serial port: {e}")
-        except Exception as e:
-            logger.error(f"Unexpected error in MessageProcessor: {e}")
-        finally:
-            if stream.in_waiting:
-                logger.debug("Data still available in stream after processing.")
-            else:
-                logger.debug("No more data available in stream.")
-        return None
 
 class SerialCommunication(CommunicationInterface):
     def __init__(self, port, baudrate, timeout, device_id, experiment_id, gnss_messages, event_bus):
