+++ src/handlers/message_handler.py
@@ -1,0 +1,6 @@
+from abc import ABC, abstractmethod
+
+class MessageHandler(ABC):
+    @abstractmethod
+    def process(self, parsed_data, device_id, experiment_id):
+        pass
